requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
Faker==12.3.3
pydantic>=1.9.0
pytest>=7.0.1
requests>=2.27.1
pytest-rerunfailures
python-dotenv
